FIX : MaxGroupCnt 5로 설정(일반 사용자)

+++ viewer4doc-admin-api/src/main/java/com/yangql/viewer4doc/interfaces/FileController.java
@@ -99,9 +99,9 @@ public class FileController {
     ){
         Claims claims = (Claims)authentication.getPrincipal();
         Long userId = claims.get("userId",Long.class);
-//        if(!fileService.checkAdmin(userId)){
-//            return ResponseEntity.badRequest().body("Unauthorized");
-//        }
+        if(!fileService.checkAdmin(userId)){
+            return ResponseEntity.badRequest().body("Unauthorized");
+        }
         fileService.deleteAllFiles(fileId,userId);
         //shareService.deleteAllShares(fileId);
